+++ src/lib/api/ContactApi.js
@@ -9,3 +9,26 @@ export const contactCreate = async (token, { first_name, last_name, email, phone
     body: JSON.stringify({ first_name, last_name, email, phone }),
   });
 };
+
+export const contactList = async (token, { name, email, phone, page }) => {
+  const url = new URL(`${import.meta.env.VITE_URL_API}/contacts`);
+  if (name) {
+    url.searchParams.append('name', name);
+  }
+  if (email) {
+    url.searchParams.append('email', email);
+  }
+  if (phone) {
+    url.searchParams.append('phone', phone);
+  }
+  if (page) {
+    url.searchParams.append('page', page);
+  }
+  return fetch(url, {
+    method: 'GET',
+    headers: {
+      Accept: 'application/json',
+      Authorization: token,
+    },
+  });
+};